+++ packages/nostrify/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.38.0 - 2025-01-30
+
+### Added
+
+- NRelay1: added relay logging.
+
 ## 0.37.0 - 2024-12-30
 
 ### Added